feat: add decompress function
